Refactor SemanticKernelService and simplify methods
The codebase has been updated to improve the maintainability of SemanticKernelService. The service was also moved from the semantic to the service package. The refactoring includes the simplification of dispatch and response processing methods. Unused imports were removed, and the handling of Kernel building is now more streamlined and centralized.
